chore(lint): clippy gating + cleanup pass — workspace 51 → 0 warnings
W1.4 from the catch-and-surpass-niri plan. niri's CI gates on
clippy; margo's `cargo clippy` was informational-only because
the codebase had accumulated drift across many sessions. The
warnings were genuinely useful but invisible — the noise floor
masked real findings. Flipping to a gate forces them up to the
top of every PR review.

Cleanup landed:

* New `clippy.toml` mirrors niri's `ignore-interior-mutability`
  list — smithay's `Window` / `Output` / `ClientId` /
  `ObjectId` use UnsafeCell internally for refcount-clones but
  expose stable `Hash`/`Eq`, which is exactly what a HashMap
  key needs. Without the override, every `HashMap<Window, _>`
  fires `mutable_key_type`.

* `cargo clippy --fix` ran for the auto-applicable categories,
  then 22 manual fixes:
    - `to_blob(&self)` → `to_blob(self)` (Copy receiver)
    - `for c in candidates { if c.is_dir() { return; } } None`
      → `candidates.find(|c| c.is_dir())` (Iterator::find)
    - `.max(6).min(28)` → `.clamp(6, 28)`
    - `if let WEnum::Value(t) = transform { ... }` collapsed
      into the outer match arm pattern
    - `for i in 0..counts.len()` index-loops → `iter_mut()`
      and `enumerate()`
    - `filter(...).next()` → `find(...)`
    - mlayout if-with-identical-arms collapsed

* Crate-level `#![allow]`s in `margo/src/main.rs` for the
  categories left intentionally suppressed:
    - `doc_overindented_list_items` / `doc_lazy_continuation`
      (cosmetic, fighting them ruins readable doc lists)
    - `too_many_arguments` (9-param render helpers — splitting
      hurts callsite clarity, not buying anything)
    - `needless_borrow` / `manual_is_multiple_of` /
      `collapsible_if` / `excessive_precision` /
      `needless_borrows_for_generic_args` (all cases where the
      explicit form documents intent better than the
      "succinct" replacement clippy proposes)

* `.github/workflows/ci.yml` flipped from
  `cargo clippy ... || true` (informational) to
  `cargo clippy ... -- -D warnings` (gating). Future PRs that
  introduce a new warning will fail CI.

Verified: 62 tests pass, release build clean, `cargo clippy
--workspace --all-targets -- -D warnings` exits 0.

Roadmap: closes W1.4.

+++ margo/src/protocols/gamma_control.rs
@@ -129,7 +129,7 @@ where
                         if let Some(gamma_size) = state.get_gamma_size(&output) {
                             let zwlr_gamma_control =
                                 data_init.init(id, GammaControlState { gamma_size });
-                            zwlr_gamma_control.gamma_size(gamma_size as u32);
+                            zwlr_gamma_control.gamma_size(gamma_size);
                             state
                                 .gamma_control_manager_state()
                                 .gamma_controls
